Allow changing password from MyGames view
