+++ client/src/views/helper/components/PasswordManagement.tsx
@@ -1,44 +1,26 @@
 import React, { ReactElement, ChangeEvent, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import styled, { css } from 'styled-components';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import {
-  getUserBySerial,
-  getUserBySerialOrUsername,
-  updateUserPassword,
-} from 'client/services/userServices';
+import { getUserBySerialOrUsername } from 'client/services/userServices';
 import { Button } from 'client/components/Button';
-import { passwordLength } from 'client/utils/validate';
+import { ChangePasswordForm } from 'client/views/helper/components/ChangePasswordForm';
 
-interface Props {
-  allowUsernameSearch: boolean;
-}
-
-export const PasswordManagement = ({
-  allowUsernameSearch = false,
-}: Props): ReactElement => {
+export const PasswordManagement = (): ReactElement => {
   const { t } = useTranslation();
 
   const [serial, setSerial] = useState<string>('');
   const [username, setUsername] = useState<string>('');
-  const [passwordFieldType, setPasswordFieldType] =
-    useState<string>('password');
   const [userSerialInput, setUserSerialInput] = useState<string>('');
-  const [changePasswordInput, setChangePasswordInput] = useState<string>('');
   const [changePasswordInputVisible, setChangePasswordInputVisible] =
     useState<boolean>(false);
   const [userFoundMessage, setUserFoundMessage] = useState<ReactElement>(
     <Message />
   );
-  const [passwordChangeMessage, setPasswordChangeMessage] =
-    useState<ReactElement>(<Message />);
 
   const submitGetUser = async (): Promise<void> => {
     if (userSerialInput.length === 0) return;
 
-    const response = allowUsernameSearch
-      ? await getUserBySerialOrUsername(userSerialInput)
-      : await getUserBySerial(userSerialInput);
+    const response = await getUserBySerialOrUsername(userSerialInput);
 
     if (!response || response.status === 'error') {
       setUserFoundMessage(
@@ -58,64 +40,18 @@ export const PasswordManagement = ({
     }
   };
 
-  const submitUpdatePassword = async (): Promise<void> => {
-    const validationError = passwordLength(changePasswordInput);
-
-    if (validationError) {
-      setPasswordChangeMessage(
-        <Message error={true}>{t(validationError)}</Message>
-      );
-      return;
-    }
-
-    const response = await updateUserPassword(
-      username,
-      serial,
-      changePasswordInput,
-      true
-    );
-    if (!response || response.status === 'error') {
-      setPasswordChangeMessage(
-        <Message error={true}>
-          {t('passwordManagement.changingPasswordError')}
-        </Message>
-      );
-    } else if (response.status === 'success') {
-      setPasswordChangeMessage(
-        <Message>{t('passwordManagement.changingPasswordSuccess')}</Message>
-      );
-    }
-  };
-
   const handleSerialChange = (event: ChangeEvent<HTMLInputElement>): void => {
     setUserSerialInput(event.target.value);
-  };
-
-  const handlePasswordChange = (event: ChangeEvent<HTMLInputElement>): void => {
-    setChangePasswordInput(event.target.value);
-  };
-
-  const togglePasswordVisibility = (): void => {
-    if (passwordFieldType === 'password') setPasswordFieldType('text');
-    else if (passwordFieldType === 'text') setPasswordFieldType('password');
   };
 
   return (
     <div className='password-management'>
       <h3>{t('passwordManagement.helperPasswordManagement')}</h3>
       <div>
-        <p>
-          {allowUsernameSearch
-            ? t('passwordManagement.userCodeOrUsername')
-            : t('passwordManagement.userCode')}
-        </p>
+        <p>{t('passwordManagement.userCodeOrUsername')}</p>
         <FormInput
           key='user-serial'
-          placeholder={
-            allowUsernameSearch
-              ? t('passwordManagement.userCodeOrUsername')
-              : t('passwordManagement.userCode')
-          }
+          placeholder={t('passwordManagement.userCodeOrUsername')}
           value={userSerialInput}
           onChange={handleSerialChange}
         />
@@ -126,36 +62,12 @@ export const PasswordManagement = ({
 
       <div>
         {changePasswordInputVisible && (
-          <>
-            <p>{t('passwordManagement.newPassword')}</p>
-            <FormInput
-              type={passwordFieldType}
-              key='new-password'
-              placeholder={t('passwordManagement.newPassword')}
-              value={changePasswordInput}
-              onChange={handlePasswordChange}
-            />
-            <Button onClick={submitUpdatePassword}>{t('button.save')}</Button>
-
-            <FormFieldIcon>
-              <FontAwesomeIcon
-                icon={passwordFieldType === 'password' ? 'eye' : 'eye-slash'}
-                onClick={togglePasswordVisibility}
-              />
-            </FormFieldIcon>
-
-            {passwordChangeMessage}
-          </>
+          <ChangePasswordForm serial={serial} username={username} />
         )}
       </div>
     </div>
   );
 };
-
-const FormFieldIcon = styled.span`
-  padding: 0 0 0 8px;
-  font-size: ${(props) => props.theme.iconSize};
-`;
 
 const FormInput = styled.input`
   border: 1px solid ${(props) => props.theme.borderInactive};
