+++ client/src/views/login/components/LoginForm.tsx
@@ -1,17 +1,14 @@
-import React, { ReactElement, useState } from 'react';
+import React, { ReactElement } from 'react';
 import { Field, reduxForm, InjectedFormProps } from 'redux-form';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 import { required } from 'client/utils/validate';
 import { FormField } from 'client/components/FormField';
 import { Button } from 'client/components/Button';
-import { PasswordManagement } from 'client/views/helper/components/PasswordManagement';
-import { Paragraph } from 'client/components/Paragraph';
 
 const LoginForm = (props: InjectedFormProps): ReactElement => {
   const { handleSubmit, submitting, error } = props;
   const { t } = useTranslation();
-  const [showChangePassword, setShowChangePassword] = useState<boolean>(false);
 
   return (
     <div className='login-form'>
@@ -40,19 +37,7 @@ const LoginForm = (props: InjectedFormProps): ReactElement => {
         <ErrorMessage>{error}</ErrorMessage>
       )}
 
-      <Button
-        selected={showChangePassword}
-        onClick={() => setShowChangePassword(!showChangePassword)}
-      >
-        {t('login.forgotPassword')}
-      </Button>
-
-      {showChangePassword && (
-        <>
-          <Paragraph text={t('passwordManagement.passwordChangeHelp')} />
-          <PasswordManagement allowUsernameSearch={false} />
-        </>
-      )}
+      <p>{t('login.passwordHint')}</p>
     </div>
   );
 };
